Some additions to previous patch + downmerge from 1.1
The merge was done manually copying relevant code changes from 1.1.
In particulary merged the patch for CONCPP-94
Also many changes, that fix real possible issues, that was also present in 1.0 and found while
working on 1.1. For example(but not limited to), there were many cases of incorrect
find_first_of std::string method, where find had to be really used.

+++ src/com/CmdInformationBatch.cpp
@@ -199,7 +199,7 @@ namespace mariadb
       if (updateCount != Statement::EXECUTE_FAILED
         && updateCount != RESULT_SET_VALUE
         && (insertId= *idIterator) > 0) {
-        for (int32_t i= 0; i < updateCount; i++) {
+        for (int64_t i= 0; i < updateCount; i++) {
           ret[position++]= insertId + i*autoIncrement;
         }
       }
